add solver into main.py
