+++ input_handling/core/lexer.py
@@ -1,10 +1,12 @@
-from core.tokens import Token, TokenType
 import string
 import os
 import sys
 
 path = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
 sys.path.insert(0, path)
+
+from core.tokens import Token, TokenType
+
 
 WHITESPACE = ' \n\t\'\"'
 DIGITS = '0123456789'
