+++ input_handling/test/tempCodeRunnerFile.py
@@ -1,0 +1,1 @@
+(x+1)^2 - 3*x
